Recover old file structure for now
