+++ aquadoggo/src/graphql/client/mod.rs
@@ -1,7 +1,9 @@
 // SPDX-License-Identifier: AGPL-3.0-or-later
 
+mod request;
+mod response;
 mod root;
-mod types;
 
+pub use request::{EntryArgsRequest, PublishEntryRequest};
+pub use response::{EntryArgsResponse, PublishEntryResponse};
 pub use root::ClientRoot;
-pub use types::{EntryArgsRequest, EntryArgsResponse, PublishEntryRequest, PublishEntryResponse};
